+++ src/madgwick.cpp
@@ -43,7 +43,7 @@ Madgwick::Madgwick() {
 /** \brief Constructor using custom sample frequency.
  * \param sample_frequency The anticipated update frequency in Hz.
  */
-Madgwick::Madgwick(float sample_frequency)
+Madgwick::Madgwick(double sample_frequency)
 {
 	beta = betaDef;
 	q0 = 1.0f;
@@ -67,12 +67,12 @@ Madgwick::Madgwick(float sample_frequency)
  *  \param mz z-axis magnetfic field strength
  *  \return Nothing; updates internal state.
  */
-void Madgwick::update(float gx, float gy, float gz, float ax, float ay, float az, float mx, float my, float mz) {
-	float recipNorm;
-	float s0, s1, s2, s3;
-	float qDot1, qDot2, qDot3, qDot4;
-	float hx, hy;
-	float _2q0mx, _2q0my, _2q0mz, _2q1mx, _2bx, _2bz, _4bx, _4bz, _2q0, _2q1, _2q2, _2q3, _2q0q2, _2q2q3, q0q0, q0q1, q0q2, q0q3, q1q1, q1q2, q1q3, q2q2, q2q3, q3q3;
+void Madgwick::update(double gx, double gy, double gz, double ax, double ay, double az, double mx, double my, double mz) {
+	double recipNorm;
+	double s0, s1, s2, s3;
+	double qDot1, qDot2, qDot3, qDot4;
+	double hx, hy;
+	double _2q0mx, _2q0my, _2q0mz, _2q1mx, _2bx, _2bz, _4bx, _4bz, _2q0, _2q1, _2q2, _2q3, _2q0q2, _2q2q3, q0q0, q0q1, q0q2, q0q3, q1q1, q1q2, q1q3, q2q2, q2q3, q3q3;
 
 	// Use IMU algorithm if magnetometer measurement invalid (avoids NaN in magnetometer normalisation)
 	if((mx == 0.0f) && (my == 0.0f) && (mz == 0.0f)) {
@@ -179,7 +179,7 @@ void Madgwick::update(float gx, float gy, float gz, float ax, float ay, float az
  *  \param az x-axis accelerometer value in m/s2
  *  \return Nothing; updates internal state.
  */
-void Madgwick::updateIMU(double timestamp, float gx, float gy, float gz, float ax, float ay, float az)
+void Madgwick::updateIMU(double timestamp, double gx, double gy, double gz, double ax, double ay, double az)
 {
 	updateIMU(gx, gy, gz, ax, ay, az);
 }
@@ -193,11 +193,11 @@ void Madgwick::updateIMU(double timestamp, float gx, float gy, float gz, float a
  *  \param az x-axis accelerometer value in m/s2
  *  \return Nothing; updates internal state.
  */
-void Madgwick::updateIMU(float gx, float gy, float gz, float ax, float ay, float az) {
-	float recipNorm;
-	float s0, s1, s2, s3;
-	float qDot1, qDot2, qDot3, qDot4;
-	float _2q0, _2q1, _2q2, _2q3, _4q0, _4q1, _4q2 ,_8q1, _8q2, q0q0, q1q1, q2q2, q3q3;
+void Madgwick::updateIMU(double gx, double gy, double gz, double ax, double ay, double az) {
+	double recipNorm;
+	double s0, s1, s2, s3;
+	double qDot1, qDot2, qDot3, qDot4;
+	double _2q0, _2q1, _2q2, _2q3, _4q0, _4q1, _4q2 ,_8q1, _8q2, q0q0, q1q1, q2q2, q3q3;
 
 	// This factor isn't required because our sensor already operates in radians/sec.
 	// Convert gyroscope degrees/sec to radians/sec
@@ -272,12 +272,12 @@ void Madgwick::updateIMU(float gx, float gy, float gz, float ax, float ay, float
  * \param x Find square root of this number.
  * \return The square root of x.
  */
-float Madgwick::invSqrt(float x) {
-	float halfx = 0.5f * x;
-	float y = x;
+double Madgwick::invSqrt(double x) {
+	double halfx = 0.5f * x;
+	double y = x;
 	long i = *(long*)&y;
 	i = 0x5f3759df - (i>>1);
-	y = *(float*)&i;
+	y = *(double*)&i;
 	y = y * (1.5f - (halfx * y * y));
 	y = y * (1.5f - (halfx * y * y));
 	return y;
